+++ src/ServiceMap/ServiceMap.php
@@ -13,20 +13,8 @@ use DigitalCraftsman\CQSRouting\RequestDataTransformer\RequestDataTransformer;
 use DigitalCraftsman\CQSRouting\RequestDecoder\RequestDecoder;
 use DigitalCraftsman\CQSRouting\RequestValidator\RequestValidator;
 use DigitalCraftsman\CQSRouting\ResponseConstructor\ResponseConstructor;
-use DigitalCraftsman\CQSRouting\ServiceMap\Exception\ConfiguredCommandHandlerNotAvailable;
-use DigitalCraftsman\CQSRouting\ServiceMap\Exception\ConfiguredDTOConstructorNotAvailable;
-use DigitalCraftsman\CQSRouting\ServiceMap\Exception\ConfiguredDTOValidatorNotAvailable;
-use DigitalCraftsman\CQSRouting\ServiceMap\Exception\ConfiguredHandlerWrapperNotAvailable;
-use DigitalCraftsman\CQSRouting\ServiceMap\Exception\ConfiguredQueryHandlerNotAvailable;
-use DigitalCraftsman\CQSRouting\ServiceMap\Exception\ConfiguredRequestDataTransformerNotAvailable;
-use DigitalCraftsman\CQSRouting\ServiceMap\Exception\ConfiguredRequestDecoderNotAvailable;
-use DigitalCraftsman\CQSRouting\ServiceMap\Exception\ConfiguredRequestValidatorNotAvailable;
-use DigitalCraftsman\CQSRouting\ServiceMap\Exception\ConfiguredResponseConstructorNotAvailable;
-use DigitalCraftsman\CQSRouting\ServiceMap\Exception\DTOConstructorOrDefaultDTOConstructorMustBeConfigured;
-use DigitalCraftsman\CQSRouting\ServiceMap\Exception\RequestDecoderOrDefaultRequestDecoderMustBeConfigured;
-use DigitalCraftsman\CQSRouting\ServiceMap\Exception\ResponseConstructorOrDefaultResponseConstructorMustBeConfigured;
 use Psr\Container\ContainerExceptionInterface;
-use Psr\Container\NotFoundExceptionInterface;
+use Symfony\Component\DependencyInjection\Attribute\AutowireLocator;
 use Symfony\Contracts\Service\ServiceProviderInterface;
 
 /**
@@ -35,142 +23,168 @@ use Symfony\Contracts\Service\ServiceProviderInterface;
 final readonly class ServiceMap
 {
     public function __construct(
+        #[AutowireLocator('cqs_routing.request_validator')]
         private ServiceProviderInterface $requestValidators,
+        #[AutowireLocator('cqs_routing.request_decoder')]
         private ServiceProviderInterface $requestDecoders,
+        #[AutowireLocator('cqs_routing.request_data_transformer')]
         private ServiceProviderInterface $requestDataTransformers,
+        #[AutowireLocator('cqs_routing.dto_constructor')]
         private ServiceProviderInterface $dtoConstructors,
+        #[AutowireLocator('cqs_routing.dto_validator')]
         private ServiceProviderInterface $dtoValidators,
+        #[AutowireLocator('cqs_routing.handler_wrapper')]
         private ServiceProviderInterface $handlerWrappers,
+        #[AutowireLocator('cqs_routing.command_handler')]
         private ServiceProviderInterface $commandHandlers,
+        #[AutowireLocator('cqs_routing.query_handler')]
         private ServiceProviderInterface $queryHandlers,
+        #[AutowireLocator('cqs_routing.response_constructor')]
         private ServiceProviderInterface $responseConstructors,
     ) {
     }
 
     /**
-     * @param class-string<RequestValidator> $requestValidatorClass
+     * @template T of RequestValidator
+     *
+     * @param class-string<T> $requestValidatorClass
+     *
+     * @return T
      */
     public function getRequestValidator(string $requestValidatorClass): RequestValidator
     {
         try {
             return $this->requestValidators->get($requestValidatorClass);
         } catch (ContainerExceptionInterface) {
-            throw new ConfiguredRequestValidatorNotAvailable($requestValidatorClass);
+            throw new Exception\ConfiguredRequestValidatorNotAvailable($requestValidatorClass);
         }
     }
 
     /**
-     * @param class-string<RequestDecoder>|null $requestDecoderClass
-     * @param class-string<RequestDecoder>|null $defaultRequestDecoderClass
+     * @template T of RequestDecoder
+     *
+     * @param class-string<T> $requestDecoderClass
+     *
+     * @return T
      */
-    public function getRequestDecoder(?string $requestDecoderClass, ?string $defaultRequestDecoderClass): RequestDecoder
+    public function getRequestDecoder(string $requestDecoderClass): RequestDecoder
     {
-        $selectedRequestDecoderClass = $requestDecoderClass ?? $defaultRequestDecoderClass;
-        if ($selectedRequestDecoderClass === null) {
-            throw new RequestDecoderOrDefaultRequestDecoderMustBeConfigured();
-        }
-
         try {
-            return $this->requestDecoders->get($selectedRequestDecoderClass);
-        } catch (NotFoundExceptionInterface) {
-            throw new ConfiguredRequestDecoderNotAvailable($selectedRequestDecoderClass);
+            return $this->requestDecoders->get($requestDecoderClass);
+        } catch (ContainerExceptionInterface) {
+            throw new Exception\ConfiguredRequestDecoderNotAvailable($requestDecoderClass);
         }
     }
 
     /**
-     * @param class-string<RequestDataTransformer> $requestDataTransformerClass
+     * @template T of RequestDataTransformer
+     *
+     * @param class-string<T> $requestDataTransformerClass
+     *
+     * @return T
      */
     public function getRequestDataTransformer(string $requestDataTransformerClass): RequestDataTransformer
     {
         try {
             return $this->requestDataTransformers->get($requestDataTransformerClass);
         } catch (ContainerExceptionInterface) {
-            throw new ConfiguredRequestDataTransformerNotAvailable($requestDataTransformerClass);
+            throw new Exception\ConfiguredRequestDataTransformerNotAvailable($requestDataTransformerClass);
         }
     }
 
     /**
-     * @param class-string<DTOConstructor>|null $dtoConstructorClass
-     * @param class-string<DTOConstructor>|null $defaultDTOConstructorClass
+     * @template T of DTOConstructor
+     *
+     * @param class-string<T> $dtoConstructorClass
+     *
+     * @return T
      */
-    public function getDTOConstructor(?string $dtoConstructorClass, ?string $defaultDTOConstructorClass): DTOConstructor
+    public function getDTOConstructor(string $dtoConstructorClass): DTOConstructor
     {
-        $selectedDTOConstructorClass = $dtoConstructorClass ?? $defaultDTOConstructorClass;
-        if ($selectedDTOConstructorClass === null) {
-            throw new DTOConstructorOrDefaultDTOConstructorMustBeConfigured();
-        }
-
         try {
-            return $this->dtoConstructors->get($selectedDTOConstructorClass);
+            return $this->dtoConstructors->get($dtoConstructorClass);
         } catch (ContainerExceptionInterface) {
-            throw new ConfiguredDTOConstructorNotAvailable($selectedDTOConstructorClass);
+            throw new Exception\ConfiguredDTOConstructorNotAvailable($dtoConstructorClass);
         }
     }
 
     /**
-     * @param class-string<DTOValidator> $dtoValidatorClass
+     * @template T of DTOValidator
+     *
+     * @param class-string<T> $dtoValidatorClass
+     *
+     * @return T
      */
     public function getDTOValidator(string $dtoValidatorClass): DTOValidator
     {
         try {
             return $this->dtoValidators->get($dtoValidatorClass);
         } catch (ContainerExceptionInterface) {
-            throw new ConfiguredDTOValidatorNotAvailable($dtoValidatorClass);
+            throw new Exception\ConfiguredDTOValidatorNotAvailable($dtoValidatorClass);
         }
     }
 
     /**
-     * @param class-string<HandlerWrapper> $handlerWrapperClass
+     * @template T of HandlerWrapper
+     *
+     * @param class-string<T> $handlerWrapperClass
+     *
+     * @return T
      */
     public function getHandlerWrapper(string $handlerWrapperClass): HandlerWrapper
     {
         try {
             return $this->handlerWrappers->get($handlerWrapperClass);
         } catch (ContainerExceptionInterface) {
-            throw new ConfiguredHandlerWrapperNotAvailable($handlerWrapperClass);
+            throw new Exception\ConfiguredHandlerWrapperNotAvailable($handlerWrapperClass);
         }
     }
 
     /**
-     * @param class-string<CommandHandler> $handlerClass
+     * @template T of CommandHandler
+     *
+     * @param class-string<T> $handlerClass
+     *
+     * @return T
      */
     public function getCommandHandler(string $handlerClass): CommandHandler
     {
         try {
             return $this->commandHandlers->get($handlerClass);
         } catch (ContainerExceptionInterface) {
-            throw new ConfiguredCommandHandlerNotAvailable($handlerClass);
+            throw new Exception\ConfiguredCommandHandlerNotAvailable($handlerClass);
         }
     }
 
     /**
-     * @param class-string<QueryHandler> $handlerClass
+     * @template T of QueryHandler
+     *
+     * @param class-string<T> $handlerClass
+     *
+     * @return T
      */
     public function getQueryHandler(string $handlerClass): QueryHandler
     {
         try {
             return $this->queryHandlers->get($handlerClass);
         } catch (ContainerExceptionInterface) {
-            throw new ConfiguredQueryHandlerNotAvailable($handlerClass);
+            throw new Exception\ConfiguredQueryHandlerNotAvailable($handlerClass);
         }
     }
 
     /**
-     * @param class-string<ResponseConstructor> $responseConstructorClass
+     * @template T of ResponseConstructor
+     *
+     * @param class-string<T> $responseConstructorClass
+     *
+     * @return T
      */
-    public function getResponseConstructor(
-        ?string $responseConstructorClass,
-        ?string $defaultResponseConstructorClass,
-    ): ResponseConstructor {
-        $selectedResponseConstructorClass = $responseConstructorClass ?? $defaultResponseConstructorClass;
-        if ($selectedResponseConstructorClass === null) {
-            throw new ResponseConstructorOrDefaultResponseConstructorMustBeConfigured();
-        }
-
+    public function getResponseConstructor(string $responseConstructorClass): ResponseConstructor
+    {
         try {
-            return $this->responseConstructors->get($selectedResponseConstructorClass);
+            return $this->responseConstructors->get($responseConstructorClass);
         } catch (ContainerExceptionInterface) {
-            throw new ConfiguredResponseConstructorNotAvailable($selectedResponseConstructorClass);
+            throw new Exception\ConfiguredResponseConstructorNotAvailable($responseConstructorClass);
         }
     }
 }