+++ src/TestUtility/Exception/RouteIsNoNotACQSRoute.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigitalCraftsman\CQSRouting\TestUtility\Exception;
+
+/**
+ * @psalm-immutable
+ *
+ * @internal
+ */
+final class RouteIsNoNotACQSRoute extends \InvalidArgumentException
+{
+    public function __construct(string $routeName)
+    {
+        parent::__construct(sprintf(
+            'The route "%s" is not a CQS route.',
+            $routeName,
+        ));
+    }
+}